+++ hunit-dejafu/CHANGELOG.rst
@@ -7,6 +7,18 @@ standard Haskell versioning scheme.
 .. _PVP: https://pvp.haskell.org/
 
 
+1.2.0.5 (2018-06-17)
+--------------------
+
+* Git: :tag:`hunit-dejafu-1.2.0.5`
+* Hackage: :hackage:`hunit-dejafu-1.2.0.5`
+
+Miscellaneous
+~~~~~~~~~~~~~
+
+* The upper bound on :hackage:`dejafu` is <1.11.
+
+
 1.2.0.4 (2018-06-10)
 --------------------